+++ packages/ripple-ui-forms/src/formkit-features.ts
@@ -1,0 +1,14 @@
+import { FormKitNode } from '@formkit/core'
+
+/**
+ * Allows us to concat two strings in formkit custom input definitions
+ */
+export const concatStrings = (node: FormKitNode): void => {
+  node.on('created', () => {
+    if (node.context?.fns) {
+      node.context.fns.concatStrings = (stringA: string, stringB: string) => {
+        return `${stringA}${stringB}`
+      }
+    }
+  })
+}